Fix substr function if length is too high
With this fix the substr function behaves more like the bash
functionality, if the string is only 5 chars long and you call `${var:0:8}` it resulted in an empty string, now it correctly returns the 5 chars. Besides that I have also added some basic tests to make sure it behaves correctly.